requirements.txt
dj-database-url==0.4.1
Django==1.10.4
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-filter==1.0.1
django-nose==1.4.4
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
Markdown==2.6.7
nose==1.3.7
Pillow==3.4.2
PyJWT==1.4.2
